+++ hapi-base/src/main/java/ca/uhn/hl7v2/util/idgenerator/InMemoryIDGenerator.java
@@ -43,7 +43,7 @@ public class InMemoryIDGenerator extends IDGenerator.OrderedSupport {
 		super();
 	}
 
-	InMemoryIDGenerator(int increment) {
+	InMemoryIDGenerator(long increment) {
 		super(increment);
 	}
