+++ hapi-base/src/main/java/ca/uhn/hl7v2/util/idgenerator/LegacyGenerator.java
@@ -48,7 +48,7 @@ public class LegacyGenerator implements IDGenerator.Ordered {
 	public void reset() {
 	}
 
-	public int getIncrement() {
+	public long getIncrement() {
 		throw new UnsupportedOperationException();
 	}
